+++ src/index.ts
@@ -2,6 +2,7 @@ import { ApplicationEndpoint } from "./endpoints/application";
 import { ApplicationRoleEndpoint } from "./endpoints/applicationRole";
 import { AttachmentEndpoint } from "./endpoints/attachment";
 import { AvatarEndpoint } from "./endpoints/avatar";
+import { ComponentEndpoint } from "./endpoints/component";
 import { PermissionsEndpoint } from "./endpoints/permissions";
 import { Basic, BasicAuth } from "./types";
 
@@ -50,9 +51,14 @@ export class JiraServerConnector {
     attachments: AttachmentEndpoint;
 
     /**
-    * Contains all operations from '/rest/api/latest/avatar/*'.
-    */
+     * Contains all operations from '/rest/api/latest/avatar/*'.
+     */
     avatar: AvatarEndpoint;
+
+    /**
+     * Contains all operations from '/rest/api/latest/component/*'.
+     */
+    component: ComponentEndpoint;
 
     /**
      * Instance new Jira Server Connector with user credentials and Jira host.
@@ -65,6 +71,7 @@ export class JiraServerConnector {
         this.applicationRoles = new ApplicationRoleEndpoint(this.auth);
         this.attachments = new AttachmentEndpoint(this.auth);
         this.avatar = new AvatarEndpoint(this.auth);
+        this.component = new ComponentEndpoint(this.auth);
     }